feat(town): extract spawnInnFollowerHires to follower_hires.js and import in runtime.js
